+++ docs/src/components/UploadKit/chains.ts
@@ -1,10 +1,10 @@
-import { bsc, mainnet } from 'wagmi/chains';
+import { bscTestnet } from 'viem/chains';
 
 export const chains = [
   {
     id: 5600,
-    name: 'BNB Greenfield Testnet',
-    network: 'BNB Greenfield Testnet',
+    name: 'BNB Greenfield Chain Testnet',
+    network: 'BNB Greenfield Chain Testnet',
     nativeCurrency: {
       name: 'tBNB',
       symbol: 'tBNB',
@@ -20,59 +20,14 @@ export const chains = [
     },
     blockExplorers: {
       etherscan: {
-        name: `BNB Greenfield Testnet Scan`,
+        name: `BNB Greenfield Chain Testnet Scan`,
         url: 'https://greenfield-chain.bnbchain.org',
       },
       default: {
-        name: `BNB Greenfield Testnet Scan`,
+        name: `BNB Greenfield Chain Testnet Scan`,
         url: 'https://greenfield-chain.bnbchain.org',
       },
     },
   },
-  {
-    id: 204,
-    name: 'opBNB Mainnet',
-    network: 'opBNB Mainnet',
-    nativeCurrency: {
-      name: 'tcBNB',
-      symbol: 'tcBNB',
-      decimals: 18,
-    },
-    rpcUrls: {
-      default: {
-        http: ['https://opbnb-mainnet-rpc.bnbchain.org'],
-      },
-      public: {
-        http: ['https://opbnb-mainnet-rpc.bnbchain.org'],
-      },
-    },
-    blockExplorers: {
-      etherscan: { name: 'opBNBScan', url: `https://mainnet.opbnbscan.com` },
-      default: { name: 'opBNBScan', url: `https://mainnet.opbnbscan.com` },
-    },
-  },
-  {
-    id: 97,
-    name: 'BSC Testnet',
-    network: 'BSC Testnet',
-    nativeCurrency: {
-      name: 'tBNB',
-      symbol: 'tBNB',
-      decimals: 18,
-    },
-    rpcUrls: {
-      default: {
-        http: [`https://data-seed-prebsc-1-s1.binance.org:8545`],
-      },
-      public: {
-        http: [`https://data-seed-prebsc-1-s1.binance.org:8545`],
-      },
-    },
-    blockExplorers: {
-      etherscan: { name: 'BSC Testnet Scan', url: `https://testnet.bscscan.com` },
-      default: { name: 'BSC Testnet Scan', url: `https://testnet.bscscan.com` },
-    },
-  },
-  bsc,
-  mainnet,
+  bscTestnet,
 ];