+++ packages/uploadkit/src/utils/error.ts
@@ -1,0 +1,8 @@
+export function catchError<T, U>(func: (...args: U[]) => T, ...args: U[]): string {
+  try {
+    func(...args);
+    return '';
+  } catch (error) {
+    return (error as Error).message;
+  }
+}